+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2020-11-05
+### Changed
+- Updated **google** and **google-beta** providers to `~> 3.0` from `~> 2.0`
+- Added a `regional` boolean variable to set the location to `var.region` if true and `var.zone` otherwise (defaults to `false`)
+
 ## [0.1.1] - 2020-08-23
 ### Changed
 - Variablized some settings